working on submit and update based on component -inventory
